backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
imagehash==4.3.1
numpy==1.25.0
scikit-learn==1.3.0
PyMuPDF==1.22.5
